requirements.txt
pandas
numpy
scikit-learn
xgboost
matplotlib
streamlit
Pillow
matplotlib